pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taskz"
version = "0.1.0"
description = "Task-Automator CLI: bulk-rename files, scrape receipts, and log expenses."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "You", email = "you@example.com" }]
license = { text = "MIT" }
dependencies = [
  "click>=8.1",
  "beautifulsoup4>=4.12",
  "lxml>=5.2",
  "python-dateutil>=2.9",
  "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "black>=24.8",
  "mypy>=1.10",
  "pre-commit>=3.8",
]
exif = ["Pillow>=10.4"]
pdf = ["pdfminer.six>=20231228"]

[project.scripts]
taskz = "taskz.cli:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = false
disallow_untyped_defs = false
show_error_codes = true

[tool.pytest.ini_options]
addopts = "-q"

[tool.hatch.build.targets.wheel]
packages = ["src/taskz"]
include = ["LICENSE", "README.md", "CHANGELOG.md"]